Applicant Profile After spending two days debugging the Applicant Profile feature, I discovered that the main issue was caused by the relationships between tables. The delete operation always failed because of related records in other tables, which prevented proper deletion. I also encountered errors while updating records, which made the process even more difficult. The fix involved enabling cascade delete in the DbContext so that related data would be removed automatically, and adjusting the update logic to work with the existing entity in the database instead of creating a new one

+++ Models/Job.cs
@@ -32,11 +32,11 @@ namespace Rezk_Proj.Models
         [Required]
         public decimal Longitude { get; set; }
 
-        [Range(0, double.MaxValue)]
-        public double MinSalary { get; set; }
+        [Range(0, float.MaxValue)]
+        public float MinSalary { get; set; }
 
-        [Range(0, double.MaxValue)]
-        public double MaxSalary { get; set; }
+        [Range(0, float.MaxValue)]
+        public float MaxSalary { get; set; }
 
 
         [Required]
